✨Add support to export the puzzle as an image
Obviously if just the text is used, it would be easy to just search
(depending on the direction potentially). Making this an image will
at least encourage cheaters to develop cool OCR solutions. I could
obfuscate the image more to make that harder, but I don't think it's
worth it.